// src/interfaces/payment.d.ts
export interface Compra {
    name:string
    description:string
    picture:string
    price:number
    nombre_comprador:string
    email_comprador:string
    codigo_de_area:string
    celular:string
    dni:string
    provincia:string
    calle:string
    numeracion:string
    codigo_postal:string
    id_producto:string | undefined
}